Seção 03 (Exercicio11 - Resolvido e Atualizado)

+++ Seção 03 (Estruturas de Controle)/Exercicios/Exercicio11/Program.cs
@@ -1,5 +1,5 @@
 ﻿Console.Write("Informe o primeiro número: ");
-int num1 = int.Parse(Console.ReadLine());
+double num1 = int.Parse(Console.ReadLine());
 
 Console.Write("Informe o operador (+,-,/,*): ");
 string operador = Console.ReadLine();
@@ -8,22 +8,22 @@ switch (operador)
 {
     case "+":
         Console.Write("Informe o segundo número: ");
-        int num2 = int.Parse(Console.ReadLine());
+        double num2 = int.Parse(Console.ReadLine());
         Console.WriteLine($"Resultado: {num1 + num2}");
         break;
     case "-":
         Console.Write("Informe o segundo número: ");
-        int num3 = int.Parse(Console.ReadLine());
+        double num3 = int.Parse(Console.ReadLine());
         Console.WriteLine($"Resultado: {num1 - num3}");
         break;
     case "*":
         Console.Write("Informe o segundo número: ");
-        int num4 = int.Parse(Console.ReadLine());
+        double num4 = int.Parse(Console.ReadLine());
         Console.WriteLine($"Resultado: {num1 * num4}");
         break;
     case "/":
         Console.Write("Informe o segundo número: ");
-        int num5 = int.Parse(Console.ReadLine());
+        double num5 = int.Parse(Console.ReadLine());
         if (num5 != 0)
         {
             Console.WriteLine($"Resultado: {num1 / num5}");
